LineageParts: Misc fixes
* Convert switch-case to if-conditions when there aren't many cases
* Use Standardcharsets
* Use try-with-resource
* Explicitly cast where required
* Use Math.min / Math.max
* Instead of storing the Context, store the ContentResolver

+++ src/org/lineageos/lineageparts/notificationlight/AlphaPatternDrawable.java
@@ -84,8 +84,8 @@ public class AlphaPatternDrawable extends Drawable {
         int height = bounds.height();
         int width = bounds.width();
 
-        numRectanglesHorizontal = (int) Math.ceil((width / mRectangleSize));
-        numRectanglesVertical = (int) Math.ceil(height / mRectangleSize);
+        numRectanglesHorizontal = (int) Math.ceil((float) width / mRectangleSize);
+        numRectanglesVertical = (int) Math.ceil((float) height / mRectangleSize);
 
         generatePatternBitmap();
     }
